# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Meeting_Chat API",
        "description": "Application for free communication",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Application for free communication"
        }
    ],
    "paths": {
        "/interlocutor/{interlocutorCode}": {
            "get": {
                "tags": [
                    "Interlocutors"
                ],
                "summary": "Show interlocutor",
                "description": "Show interlocutor",
                "operationId": "interlocutor",
                "parameters": [
                    {
                        "name": "interlocutorCode",
                        "in": "path",
                        "description": "Interlocutor code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetInterlocutor"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/interlocutor": {
            "get": {
                "tags": [
                    "Interlocutors"
                ],
                "summary": "Displays interlocutors",
                "description": "Displays interlocutors - with oneOf examples",
                "operationId": "displays_interlocutors",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "A list of things.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response_index_interlocutors"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/interlocutor/{id}": {
            "delete": {
                "tags": [
                    "Interlocutors"
                ],
                "summary": "Destroy interlocutor",
                "description": "Destroy interlocutor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter with mutliple examples",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetInterlocutor"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "message": "No query results for model [App\\Models\\Interlocutor] 20."
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/room/{roomName}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Show room",
                "description": "Show room",
                "parameters": [
                    {
                        "name": "roomName",
                        "in": "path",
                        "description": "Room name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetRoom"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/room/check/{name}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Show room",
                "description": "Show room",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Room name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "text/html": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        {
                                            "type": "string",
                                            "example": "0"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/interlocutors/{roomName}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Displays room interlocutors",
                "description": "Displays room interlocutors",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Room name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response_room_interlocutors"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/join/{roomName}/{interlocutorCode}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Join the room",
                "description": "Join the room",
                "parameters": [
                    {
                        "name": "roomName",
                        "in": "path",
                        "description": "Room name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "interlocutorCode",
                        "in": "path",
                        "description": "Interlocutor code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                }
            }
        },
        "/room/{interlocutorCode}": {
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Adds a new room",
                "description": "Adds a new room",
                "parameters": [
                    {
                        "name": "interlocutorCode",
                        "in": "path",
                        "description": "Interlocutor code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Room name",
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "dd39ff46-62c7-362b-959c-0cef01be272d"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "GetInterlocutor": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "interlocutorName": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    },
                    "user": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/GetUser"
                            }
                        ]
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "interlocutorName": "Test",
                    "code": "23984y-uioe897-829347g-34879gb23",
                    "user": "null",
                    "created_at": "2023-11-06T15:34:03.000000Z",
                    "updated_at": "2023-11-06T15:34:03.000000Z"
                }
            },
            "Response_index_interlocutors": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "interlocutorName": {
                                    "type": "string"
                                },
                                "code": {
                                    "type": "string"
                                },
                                "user": {
                                    "oneOf": [
                                        {
                                            "type": "null"
                                        },
                                        {
                                            "$ref": "#/components/schemas/GetUser"
                                        }
                                    ]
                                },
                                "created_at": {
                                    "type": "string"
                                },
                                "updated_at": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "links": {
                        "type": "object",
                        "allOf": [
                            {
                                "properties": {
                                    "path": {
                                        "type": "string"
                                    },
                                    "firstPageUrl": {
                                        "type": "string"
                                    },
                                    "lastPageUrl": {
                                        "type": "string"
                                    },
                                    "nextPageUrl": {
                                        "type": "string"
                                    },
                                    "prevPageUrl": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        ]
                    },
                    "meta": {
                        "type": "object",
                        "allOf": [
                            {
                                "properties": {
                                    "currentPage": {
                                        "type": "integer"
                                    },
                                    "from": {
                                        "type": "integer"
                                    },
                                    "lastPage": {
                                        "type": "integer"
                                    },
                                    "perPage": {
                                        "type": "integer"
                                    },
                                    "to": {
                                        "type": "integer"
                                    },
                                    "total": {
                                        "type": "integer"
                                    },
                                    "count": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        ]
                    }
                },
                "type": "object",
                "example": {
                    "data": [
                        {
                            "id": 1,
                            "interlocutorName": "Test",
                            "code": "23984y-uioe897-829347g-34879gb23",
                            "user": "null",
                            "created_at": "2023-11-06T15:34:03.000000Z",
                            "updated_at": "2023-11-06T15:34:03.000000Z"
                        },
                        {
                            "id": 2,
                            "interlocutorName": "Test",
                            "code": "23984y-uioe897-829347g-34879gb23",
                            "user": "null",
                            "created_at": "2023-11-06T15:34:03.000000Z",
                            "updated_at": "2023-11-06T15:34:03.000000Z"
                        }
                    ],
                    "links": {
                        "path": "http://localhost:8000/api/interlocutor",
                        "firstPageUrl": "http://localhost:8000/api/interlocutor?page=1",
                        "lastPageUrl": "http://localhost:8000/api/interlocutor?page=2",
                        "nextPageUrl": "http://localhost:8000/api/interlocutor?page=2",
                        "prevPageUrl": null
                    },
                    "meta": {
                        "currentPage": 1,
                        "from": 1,
                        "lastPage": 2,
                        "perPage": 10,
                        "to": 10,
                        "total": 13,
                        "count": 10
                    }
                }
            },
            "GetRoom": {
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "creator": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "name": "23984y-uioe897-829347g-34879gb23",
                    "creator": 2,
                    "created_at": "2023-11-06T15:34:03.000000Z",
                    "updated_at": "2023-11-06T15:34:03.000000Z"
                }
            },
            "Response_room_interlocutors": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/GetInterlocutor"
                }
            },
...
            "GetUser": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "currentAddress": {
                        "type": "string"
                    },
                    "permanantAddress": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "birthdey": {
                        "type": "string"
                    },
                    "email_verified": {
                        "type": "string"
                    },
                    "avatarPath": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "firstName": "Test",
                    "lastName": "User",
                    "gender": "Female",
                    "phone": "+380501234567",
                    "currentAddress": "Beech Creek, PA, Pennsylvania",
                    "permanantAddress": "Arlington Heights, IL, Illinois",
                    "email": "12345@gmail.com",
                    "birthday": "1975-05-21",
                    "email_verified_at": "2023-11-06T15:34:03.000000Z",
                    "avatarPath": "avatar/127486g-2397y4-28397-2387h4.png",
                    "email_verified": false,
                    "created_at": "2023-11-06T15:34:03.000000Z",
                    "updated_at": "2023-11-06T15:34:03.000000Z"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}